controllers: return empty array if no rows found, models: support float cost

+++ models/expense.go
@@ -13,7 +13,7 @@ import (
 type Expense struct {
 	Id         int       `orm:"column(id);auto" json:"id"`
 	Title      string    `orm:"column(title)" json:"title"`
-	Cost       int       `orm:"column(cost)" json:"cost"`
+	Cost       float64   `orm:"column(cost)" json:"cost"`
 	Class      int       `orm:"column(class);null" json:"class"`
 	SubClass   int       `orm:"column(sub_class);null" json:"sub_class"`
 	CreateTime time.Time `orm:"column(create_time);type(datetime);null" json:"create_time"`
